Modify logic to support uploadingURLs as an option
The Ooyala asset ingestion process requires the developer to ingest to a
set of uploading urls. Meaning, every shingle chunk needs to be uploaded
to a different url containing unique parameters.

Add support to uploadingURLs option parameter which is an array of
Ooyala API uploading URLs form which the start byte and the end byte of
the chunk will be determined from. Use a regular expression to get the
0000xxx-0000xxx start-end bytes from the URL.

Add options to the ResumableChunk Object so that the targetURL and the
start and endByte can be specifed.